IS601/loggingCalc/requirements.txt
astroid==3.3.8
coverage==7.6.10
dill==0.3.9
exceptiongroup==1.2.0
Faker==36.1.0
iniconfig==2.0.0
isort==6.0.0
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.4
pylint-pytest==1.1.8
pytest==8.2.0
pytest-cov==6.0.0
pytest-pylint==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.9.0
tzdata==2025.1
